e2e_tests/requirements.txt
attrs==21.2.0
gitdb==4.0.9
GitPython==3.1.24
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
smmap==5.0.0
toml==0.10.2
typing-extensions==4.0.0
